changed everything...need to implement teller2 and teller3
Added Q thread and the Q, added teller1 thread.  Not really using good OO programming concepts but C++ is nice because we get the queues without having to roll our own.  Need to implement teller 2 and teller 3 threads, those would be the same as teller1 but with some minor changes (variable names change to t2 and t3...).  If you really wanna utilize c++ we should make the tellers objects and use them that way.  Not sure how that'd impact threading so may be easier to just do it this way.